aetherion/pyproject.toml
[build-system]
requires = [
    "scikit-build-core",
    "nanobind>=2.7.0",
    "numpy>=2.2.4",
]

build-backend = "scikit_build_core.build"

[project]
name = "aetherion"
version = "0.1.0"
description = "Life Simulation Core module built with nanobind"
requires-python = ">=3.12"
authors = [
    { name = "Arthur Moreno", email = "morenorarthur@gmail.com" }
]
dependencies = [
  "PySDL2>=0.9.17",
]

[project.urls]
Homepage = "https://github.com/yourusername/aetherion"

[tool.scikit-build]
build.verbose = false
cmake.args = [
  "-G", "Ninja",
  "-DPython3_FIND_DEBUG_MODE=3",
  "-DCMAKE_CXX_FLAGS=-fdiagnostics-color=always",
  "-DBUILD_WORLD_TEST=OFF",
]

build-dir = "build/{wheel_tag}"

# Build stable ABI wheels for CPython 3.12+
wheel.py-api = "cp312"


[tool.setuptools.packages.find]
where = ["."]
include = ["aetherion"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
# This tells setuptools to include the header file from the install tree
aetherion = [  
  "*.pyi",      # your generated stub
  "py.typed",   # PEP 561 marker
  "include/aetherion/aetherion.hpp"
]
"" = ["*.hpp", "*.h"]